requirements.txt
Flask==2.0.2
keras==2.8.0
numpy==1.21.1
pandas==1.4.2
scikit_learn==1.1.2
scipy==1.8.0
statsmodels==0.13.2
tensorflow-cpu==2.8.0
gunicorn==20.1.0
protobuf==3.20.*